requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
schedule==1.2.0

